+++ CHANGELOG.rst
@@ -3,6 +3,14 @@
 
 ### unreleased
 
+- Add a pipeline `scan_package.ScanPackage` to scan a single package archive with
+  ScanCode-toolkit.
+  https://github.com/nexB/scancode.io/issues/25
+
+- Detected Package dependencies are not created as Package instance anymore but stored
+  on the Package model itself in a new `dependencies` field.
+  https://github.com/nexB/scancode.io/issues/228
+
 - Add the extra_data field on the DiscoveredPackage model.
   https://github.com/nexB/scancode.io/issues/191